Rename: 모든 테스트 기능 함수 LottoTest.js 로 이동
